Move bedrock.mozorg.templatetags.misc to the firefox app
Moving them to the base app caused test failures that were very hard to
debug, but were likely related to the patching of the Jinja env in various
tests, and then a lack of cleanup. We should revisit this and deal with it
