+++ bedrock/firefox/templatetags/misc.py
